+++ apps/server/src/features/news/news.controller.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { NewsController } from './news.controller';
+import { NewsController } from './controller/news.controller';
 import { NewsService } from './news.service';
-import { NewsRepo } from './repos/news.repo';
+import { NewsRepo } from './repo/news.repo';
 
 describe('NewsController', () => {
 	let controller: NewsController;
