+++ apps/server/src/models/repo/index.ts
@@ -1,0 +1,4 @@
+export interface PaginationModel {
+	readonly skip?: number;
+	readonly limit?: number;
+}
