+++ apps/server/src/modules/authentication/auth.controller.spec.ts
@@ -1,7 +1,7 @@
 import { JwtService } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { Test, TestingModule } from '@nestjs/testing';
-import { UsersModule } from '../users/users.module';
+import { UsersModule } from '../../features/users/users.module';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { jwtOptionsProvider } from './jwt-options.provider';
