+++ apps/server/src/modules/authentication/auth.service.ts
@@ -1,5 +1,5 @@
 import { Injectable, NotImplementedException } from '@nestjs/common';
-import { User, UsersService } from '../users/users.service';
+import { User, UsersService } from '../../features/users/users.service';
 import { JwtService } from '@nestjs/jwt';
 import { AuthEntity } from './entities/auth.entity';
