+++ apps/server/src/modules/authentication/dto/user.dto.ts
