+++ apps/server/src/modules/authentication/entities/auth.entity.ts
