+++ apps/server/src/modules/authentication/guards/jwt-auth.guard.ts
