+++ apps/server/src/modules/authentication/interfaces/jwt-payload.ts
