+++ apps/server/src/modules/authentication/jwt-options.provider.ts
