+++ apps/server/src/modules/authentication/strategies/jwt.strategy.ts
