+++ apps/server/src/modules/authentication/strategies/local.strategy.ts
