+++ apps/server/src/modules/authorization/feathers-service.provider.ts
